pyproject.toml
[tool.poetry]
name = "iris-podcast-maker"
version = "0.1.0"
description = "Make podcast and transcript from Iris Church from Cluj"
authors = ["candale <candale.andrei@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
yt-dlp = "^2024.12.23"
ffmpeg-python = "^0.2.0"
loguru = "^0.7.3"
pudb = "^2024.1.3"
transformers = "^4.49.0"
torch = "^2.6.0"


[tool.poetry.group.dev.dependencies]
pudb = "^2024.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
